exploration/requirements.txt
absl-py==2.1.0
ACSConv==0.1.1
aiohappyeyeballs==2.5.0
aiohttp==3.11.13
aiosignal==1.3.2
annotated-types==0.7.0
asttokens==3.0.0
astunparse==1.6.3
attrs==25.1.0
autograd==1.8.0
autokeras==2.0.0
backcall==0.2.0
beautifulsoup4==4.13.3
bleach==6.2.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
comm==0.2.2
complextorch==1.2.0
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.13
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.2.18
dm-tree==0.1.9
docker-pycreds==0.4.0
docopt==0.6.2
escnn==1.0.11
executing==2.2.0
fastjsonschema==2.21.1
filelock==3.17.0
fire==0.7.0
flatbuffers==25.2.10
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2025.3.0
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.44
google-pasta==0.2.0
grpcio==1.71.0
h5py==3.13.0
idna==3.10
imageio==2.37.0
iniconfig==2.1.0
ipykernel==6.29.5
ipython==8.12.3
ipywidgets==8.1.6
jedi==0.19.2
Jinja2==3.1.6
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
jupyterlab_widgets==3.0.14
kagglehub==0.3.10
keras==3.9.0
keras-hub==0.19.1
keras-nlp==0.19.1
keras-tuner==1.4.7
kiwisolver==1.4.8
kt-legacy==1.0.5
lazy_loader==0.4
libclang==18.1.1
lie_learn==0.0.2
lightning-utilities==0.14.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
medmnist==3.0.2
mistune==3.1.2
ml-dtypes==0.4.1
mpmath==1.3.0
multidict==6.1.0
namex==0.0.8
narwhals==1.37.1
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.2.5
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
opt_einsum==3.4.0
optree==0.14.1
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.1.0
pipreqs==0.5.0
platformdirs==4.3.6
plotly==6.0.1
pluggy==1.5.0
prettytable==3.15.1
prompt_toolkit==3.0.50
propcache==0.3.0
protobuf==5.29.3
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py3nj==0.2.1
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pymanopt==2.2.1
pyparsing==3.2.1
pytest==8.3.5
python-dateutil==2.9.0.post0
pytorch-lightning==2.5.0.post0
pytz==2025.1
PyYAML==6.0.2
pyzmq==26.2.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rpds-py==0.23.1
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
sentry-sdk==2.24.1
setproctitle==1.3.5
setuptools==76.0.0
six==1.17.0
smmap==5.0.2
soupsieve==2.6
stack-data==0.6.3
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
tensorflow==2.18.0
tensorflow-text==2.18.1
termcolor==2.5.0
threadpoolctl==3.5.0
tifffile==2025.2.18
tinycss2==1.4.0
torch==2.6.0
torch-tb-profiler==0.4.3
torchmetrics==1.6.2
torchvision==0.21.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
triton==3.2.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
wandb==0.19.8
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.1.3
wheel==0.45.1
widgetsnbextension==4.0.14
wrapt==1.17.2
yarg==0.1.9
yarl==1.18.3
